+++ fromconfig/launcher/hparams.py
@@ -7,6 +7,7 @@ import logging
 from fromconfig.core.base import fromconfig
 from fromconfig.launcher import base
 from fromconfig.utils.nest import merge_dict
+from fromconfig.utils.types import is_mapping
 
 
 LOGGER = logging.getLogger(__name__)
@@ -29,14 +30,17 @@ class HParamsLauncher(base.Launcher):
 
     def __call__(self, config: Any, command: str = ""):
         """Generate configs via hyper-parameters."""
-        hparams = fromconfig(config.get("hparams", {}))
-        if not hparams:
+        if not is_mapping(config):
             self.launcher(config=config, command=command)
         else:
-            names = hparams.keys()
-            for values in itertools.product(*[hparams[name] for name in names]):
-                overrides = dict(zip(names, values))
-                LOGGER.info("Launching with params")
-                for key, value in overrides.items():
-                    LOGGER.info(f"- {key}: {value}")
-                self.launcher(config=merge_dict(config, {"hparams": overrides}), command=command)
+            hparams = fromconfig(config.get("hparams") or {})
+            if not hparams:
+                self.launcher(config=config, command=command)
+            else:
+                names = hparams.keys()
+                for values in itertools.product(*[hparams[name] for name in names]):
+                    overrides = dict(zip(names, values))
+                    LOGGER.info("Launching with params")
+                    for key, value in overrides.items():
+                        LOGGER.info(f"- {key}: {value}")
+                    self.launcher(config=merge_dict(config, {"hparams": overrides}), command=command)