+++ fromconfig/utils/testing.py
@@ -4,6 +4,6 @@
 def assert_launcher_is_discovered(name: str, expected):
     """Test that a launcher is found by name."""
     # pylint: disable=import-outside-toplevel,cyclic-import
-    from fromconfig.launcher.base import _get_cls
+    from fromconfig.launcher.base import _classes
 
-    assert _get_cls(name) is expected
+    assert _classes()[name] is expected